More resilience  with older libcurl versions (#4)
Explicitly set the HTTP protocol version in libcurl since curl tends to change defaults in new versions.
TLS and HTTP/2 are default in newer versions. But they are not needed when talking to the Lambda runtime API.
see the defaults section in https://curl.haxx.se/libcurl/c/CURLOPT_HTTP_VERSION.html

Some older versions of libcurl (7.35.0) fail to send the request /next after the sandbox is resumed.
So, adding a retry with a max of 3 to work around that issue.
Newer versions of libcurl do _not_ have that problem and no retries are performed.