Use relative paths for CMake package config files
Using relative paths in the 'install' command results in a CMake
target file that does not contain hard-coded/machine dependent absolute
paths to the binary library file. This is useful when building and
installing the library on one machine then copying the installed files
to be consumed on another machine. A common practice when using
distributed builds.