Reduce bloat in the output zip package
Avoid dereferencing symlinks when iterating over libc files. Otherwise,
we end up with two copies of the same file, one with the symlink name
and the actual file.

Files which are directly linked to the executable are listed (in the
ldd output) by their symlink name. Those are fine to dereference
because there's only a single mention of them. The actual (dereferenced)
file is not listed in output from ldd.