Fix off-by-one bug when rpm is installed on non-RHEL systems
In 02c35e4 we removed the --quiet flag because it printed nothing on
RHEL systems regardless if the package existed or not. However, on
Debian systems without --quiet we get at least a single line if the
package is not installed "package glibc is not installed".
This patch fixes this off by one error and fixes issue #41

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.