Add gtest amalgamated source
relying on the linux distro package manager to have google-test and
installing it correctly turned out to be less than ideal. Namely Ubuntu
Bionic (18.04) failed to install the DSO and CMake 3.14.4 failed to find
it.

This is more reliable and pins the test framework too. The code will not
be part of any function that gets deployed it's merely for running the
tests