Reduce string copying.
Optimized a few unnecessary string copies that should have been moves to
begin with. Additionally, I noticed that we're scanning the headers
twice once to test if a given header exists and another to get its
value, that is a perfect use case for std::optional, but since we need
to stay compatible with C++11 and C++14, we can use the 'outcome' class
instead.